feat: load test store endpoints
