+++ bin/stress-test/Cargo.toml
@@ -16,8 +16,8 @@ version.workspace      = true
 workspace = true
 
 [dependencies]
-anyhow = { workspace = true }
-clap = { version = "4.5", features = ["derive", "string"] }
+clap = { version = "4.5", features = ["derive"] }
+futures = { version = "0.3" }
 miden-air = { workspace = true }
 miden-block-prover = { git = "https://github.com/0xPolygonMiden/miden-base.git", branch = "next", features = [
   "testing",
@@ -28,8 +28,8 @@ miden-node-proto = { workspace = true }
 miden-node-store = { workspace = true }
 miden-node-utils = { workspace = true }
 miden-objects = { workspace = true, features = ["testing"] }
-rand = { version = "0.9" }
+rand = { workspace = true }
 rayon = { version = "1.5" }
-tokio = { workspace = true, features = ["macros", "net", "rt-multi-thread", "sync", "time"] }
+tokio = { workspace = true }
 tonic = { workspace = true }
 winterfell = { version = "0.12" }
